Updated SaveXml() to reflect changes to my variant of TiXml (namely the addition of TiXmlFileInterface).
Updates to Editor: testing map saving, placing entities, misc.

Slightly modified the way Renderer works (it no longer stores Entity references.)

Fixed bug that caused sprite resources to be updated multiple times per step - one time for each Renderable that held a reference to them.

Added doubleclick catching and pointer position for "Button" type events to the RawInput signal in InputManager.

+++ FusionEngine/FusionLogPhysFS.cpp
@@ -108,7 +108,7 @@ namespace FusionEngine
 				std::string buffer("");
 				if (PHYSFS_seek(reread, seekPosition) != 0)
 				{
-					buffer.resize(keptLength, '\0');
+					buffer.resize((size_t)keptLength, '\0');
 					count = PHYSFS_read(reread, static_cast<void*>(&buffer[0]), buffer.size(), 1);
 				}
